{"schema":"decodex/commit/1","summary":"Modularize rsnap host ffi ABI","authority":"manual"}
