+++ packages/rsnap-host-ffi/src/lib.rs
@@ -4,6 +4,27 @@
 //! new host/core direction with an opaque session handle, FFI-safe config/event
 //! structs, and copy-out scene/request snapshots.
 
+mod abi;
+
+#[cfg(target_os = "macos")]
+pub use abi::RsnapLiveSamplerHandle;
+pub use abi::{
+	RSNAP_HOST_FFI_ABI_VERSION, RsnapCaptureFrameBackgroundKind, RsnapCaptureFrameBackgroundPlan,
+	RsnapCaptureFrameColorStop, RsnapCaptureFramePlan, RsnapCaptureFrameRenderKind,
+	RsnapCaptureFrameShadow, RsnapCaptureFrameSourceKind, RsnapCaptureFrameWallpaperRequest,
+	RsnapCursorIntent, RsnapFloatPoint, RsnapFloatRect, RsnapFrozenAnnotationColor,
+	RsnapFrozenOverlayEditSessionHandle, RsnapFrozenOverlayEditSnapshot,
+	RsnapFrozenOverlayEditStyle, RsnapFrozenOverlayExportElement,
+	RsnapFrozenOverlayExportElementKind, RsnapFrozenSelectionTransformKind, RsnapHostEffectKind,
+	RsnapHostEvent, RsnapHostEventKind, RsnapHostReport, RsnapHostReportKind, RsnapHostRequestKind,
+	RsnapHostRequestValue, RsnapLiveSample, RsnapMonitorRect, RsnapOwnedBytes,
+	RsnapOwnedRgbaRegion, RsnapPermissionKind, RsnapPixelRect, RsnapPlatformTag, RsnapPoint,
+	RsnapRect, RsnapRgb, RsnapRgbaRegion, RsnapSceneKind, RsnapSceneModel, RsnapScrollMinimapPlan,
+	RsnapScrollObserveOutcomeKind, RsnapScrollObserveResult, RsnapScrollSessionHandle,
+	RsnapSessionConfig, RsnapSessionHandle, RsnapStatus, RsnapToolbarItem, RsnapToolbarItemKind,
+	RsnapWindowRect,
+};
+
 use std::ffi::{CStr, CString};
 use std::mem;
 use std::os::raw::c_char;
@@ -13,6 +34,9 @@ use std::slice;
 #[cfg(not(target_os = "macos"))]
 use rsnap_overlay as _;
 
+use self::abi::{
+	RSNAP_LIVE_SAMPLE_PATCH_CAPACITY, RSNAP_STATUS_MESSAGE_CAPACITY, RSNAP_TOOLBAR_ITEM_CAPACITY,
+};
 use rsnap_capture_core::SceneModel;
 use rsnap_capture_core::{
 	self, AutoCenterImageError, BgraFrameView, CaptureFrameBackgroundKind,
@@ -42,938 +66,6 @@ use rsnap_overlay::host_live_sampling_macos::HostMacLiveSampler;
 use rsnap_overlay::scroll_stitching::{
 	ScrollStitchImage, ScrollStitchObserveOutcome, ScrollStitchSession,
 };
-
-/// ABI version exported by the thin C host bridge.
-pub const RSNAP_HOST_FFI_ABI_VERSION: u32 = 36;
-
-const RSNAP_TOOLBAR_ITEM_CAPACITY: usize = 16;
-const RSNAP_STATUS_MESSAGE_CAPACITY: usize = 256;
-const RSNAP_LIVE_SAMPLE_PATCH_CAPACITY: usize = 4_096;
-
-/// Opaque session handle owned by the native host through the C ABI.
-pub struct RsnapSessionHandle {
-	session: CaptureSessionCore,
-}
-
-/// Opaque scroll-capture stitching handle owned by the native host through the C ABI.
-pub struct RsnapScrollSessionHandle {
-	session: ScrollStitchSession,
-}
-
-/// Opaque frozen-overlay edit handle owned by the native host through the C ABI.
-pub struct RsnapFrozenOverlayEditSessionHandle {
-	session: FrozenOverlayEditSession,
-}
-
-#[cfg(target_os = "macos")]
-/// Opaque live-sampler handle owned by the native host through the C ABI.
-pub struct RsnapLiveSamplerHandle {
-	sampler: HostMacLiveSampler,
-}
-
-/// Result code returned by FFI entry points.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapStatus {
-	/// The operation succeeded.
-	Ok = 0,
-	/// The provided session handle was null.
-	NullHandle = 1,
-	/// The provided output pointer was null.
-	NullOutput = 2,
-	/// No queued value was available.
-	Empty = 3,
-	/// The provided input payload was invalid.
-	InvalidInput = 4,
-}
-
-/// FFI-safe live cursor sample copied out of the native Rust sampler.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapLiveSample {
-	/// Sampled RGB value.
-	pub rgb: RsnapRgb,
-	/// Non-zero when `rgb` is present.
-	pub has_rgb: u8,
-	/// Non-zero when frame provenance fields are present.
-	pub has_frame_metadata: u8,
-	/// Age of the sampled ScreenCaptureKit frame in microseconds.
-	pub frame_age_micros: u64,
-	/// Monotonic sequence of the sampled ScreenCaptureKit frame.
-	pub frame_seq: u64,
-	/// Live stream generation that produced the sampled frame.
-	pub stream_generation: u64,
-	/// Sampled loupe patch width in pixels.
-	pub patch_width: u32,
-	/// Sampled loupe patch height in pixels.
-	pub patch_height: u32,
-	/// Byte count copied into `patch_rgba`.
-	pub patch_len: u32,
-	/// Optional RGBA patch bytes in row-major order.
-	pub patch_rgba: [u8; RSNAP_LIVE_SAMPLE_PATCH_CAPACITY],
-}
-impl Default for RsnapLiveSample {
-	fn default() -> Self {
-		Self {
-			rgb: RsnapRgb::default(),
-			has_rgb: 0,
-			has_frame_metadata: 0,
-			frame_age_micros: 0,
-			frame_seq: 0,
-			stream_generation: 0,
-			patch_width: 0,
-			patch_height: 0,
-			patch_len: 0,
-			patch_rgba: [0; RSNAP_LIVE_SAMPLE_PATCH_CAPACITY],
-		}
-	}
-}
-
-/// FFI-safe owned RGBA image region copied out of the cached live sampler frame.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapRgbaRegion {
-	/// Region width in pixels.
-	pub width: u32,
-	/// Region height in pixels.
-	pub height: u32,
-	/// Byte count in `rgba`.
-	pub len: usize,
-	/// Caller-provided buffer capacity in bytes.
-	pub capacity: usize,
-	/// Caller-provided RGBA byte buffer in row-major order.
-	pub rgba: *mut u8,
-}
-impl Default for RsnapRgbaRegion {
-	fn default() -> Self {
-		Self { width: 0, height: 0, len: 0, capacity: 0, rgba: ptr::null_mut() }
-	}
-}
-
-/// FFI-safe owned RGBA image region whose buffer is retained by Rust until explicitly freed.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapOwnedRgbaRegion {
-	/// Region width in pixels.
-	pub width: u32,
-	/// Region height in pixels.
-	pub height: u32,
-	/// Byte count in `rgba`.
-	pub len: usize,
-	/// Reserved buffer capacity in bytes.
-	pub capacity: usize,
-	/// Owned RGBA byte buffer in row-major order.
-	pub rgba: *mut u8,
-}
-impl Default for RsnapOwnedRgbaRegion {
-	fn default() -> Self {
-		Self { width: 0, height: 0, len: 0, capacity: 0, rgba: ptr::null_mut() }
-	}
-}
-
-/// FFI-safe owned byte buffer retained by Rust until explicitly freed.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapOwnedBytes {
-	/// Byte count in `bytes`.
-	pub len: usize,
-	/// Reserved buffer capacity in bytes.
-	pub capacity: usize,
-	/// Owned byte buffer.
-	pub bytes: *mut u8,
-}
-impl Default for RsnapOwnedBytes {
-	fn default() -> Self {
-		Self { len: 0, capacity: 0, bytes: ptr::null_mut() }
-	}
-}
-
-/// FFI-safe pixel-space rectangle.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapPixelRect {
-	/// Left coordinate in pixels.
-	pub x: u32,
-	/// Top coordinate in pixels.
-	pub y: u32,
-	/// Rectangle width in pixels.
-	pub width: u32,
-	/// Rectangle height in pixels.
-	pub height: u32,
-}
-
-/// FFI-safe display-space rectangle.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapFloatRect {
-	/// Left coordinate in display points.
-	pub x: f64,
-	/// Top coordinate in display points.
-	pub y: f64,
-	/// Rectangle width in display points.
-	pub width: f64,
-	/// Rectangle height in display points.
-	pub height: f64,
-}
-
-/// FFI-safe display-space point.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapFloatPoint {
-	/// X coordinate in display points.
-	pub x: f64,
-	/// Y coordinate in display points.
-	pub y: f64,
-}
-
-/// FFI-safe frozen annotation color discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapFrozenAnnotationColor {
-	/// White annotation color.
-	White = 0,
-	/// Yellow annotation color.
-	Yellow = 1,
-	/// Green annotation color.
-	Green = 2,
-	/// Blue annotation color.
-	Blue = 3,
-	/// Red annotation color.
-	Red = 4,
-	/// Black annotation color.
-	Black = 5,
-}
-
-/// FFI-safe frozen-overlay export element discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapFrozenOverlayExportElementKind {
-	/// Pen stroke annotation.
-	Pen = 0,
-	/// Arrow annotation.
-	Arrow = 1,
-	/// Mosaic privacy rectangle.
-	Mosaic = 2,
-	/// Spotlight annotation.
-	Spotlight = 3,
-	/// Text annotation.
-	Text = 4,
-}
-
-/// FFI-safe frozen-overlay export element.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct RsnapFrozenOverlayExportElement {
-	/// Element kind.
-	pub kind: RsnapFrozenOverlayExportElementKind,
-	/// Rectangle payload for mosaic or spotlight annotations.
-	pub rect: RsnapFloatRect,
-	/// Start point, arrow tail, or text anchor.
-	pub start: RsnapFloatPoint,
-	/// Arrow tip.
-	pub end: RsnapFloatPoint,
-	/// Optional point buffer for pen strokes.
-	pub points: *const RsnapFloatPoint,
-	/// Number of points in `points`.
-	pub points_len: usize,
-	/// Optional null-terminated UTF-8 text payload.
-	pub text: *const c_char,
-	/// Stroke width in points for pen and arrow annotations.
-	pub stroke_width_points: f64,
-	/// Border width in points for spotlight annotations.
-	pub border_width_points: f64,
-	/// Font size in points for text annotations.
-	pub font_size_points: f64,
-	/// Annotation color.
-	pub color: RsnapFrozenAnnotationColor,
-}
-
-/// FFI-safe frozen-overlay edit style payload.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct RsnapFrozenOverlayEditStyle {
-	/// Stroke width in points for pen and arrow annotations.
-	pub stroke_width_points: f64,
-	/// Stroke color for pen and arrow annotations.
-	pub stroke_color: RsnapFrozenAnnotationColor,
-	/// Border width in points for spotlight annotations.
-	pub spotlight_border_width_points: f64,
-	/// Border color for spotlight annotations.
-	pub spotlight_color: RsnapFrozenAnnotationColor,
-	/// Font size in points for text annotations.
-	pub text_font_size_points: f64,
-	/// Text color.
-	pub text_color: RsnapFrozenAnnotationColor,
-}
-impl Default for RsnapFrozenOverlayEditStyle {
-	fn default() -> Self {
-		Self {
-			stroke_width_points: 3.0,
-			stroke_color: RsnapFrozenAnnotationColor::Blue,
-			spotlight_border_width_points: 0.0,
-			spotlight_color: RsnapFrozenAnnotationColor::Blue,
-			text_font_size_points: 16.0,
-			text_color: RsnapFrozenAnnotationColor::Blue,
-		}
-	}
-}
-
-/// FFI-safe owned frozen-overlay edit snapshot.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct RsnapFrozenOverlayEditSnapshot {
-	/// Non-zero when undo is available.
-	pub can_undo: u8,
-	/// Non-zero when redo is available.
-	pub can_redo: u8,
-	/// Non-zero when selection transforms should be locked out.
-	pub keeps_frozen_selection_fixed: u8,
-	/// Non-zero when a movable annotation is being dragged.
-	pub is_moving_movable_annotation: u8,
-	/// Non-zero when any pointer interaction is active.
-	pub has_active_interaction: u8,
-	/// Owned visible committed elements.
-	pub elements: *mut RsnapFrozenOverlayExportElement,
-	/// Number of visible committed elements.
-	pub elements_len: usize,
-	/// Non-zero when `preview_pen` is present.
-	pub has_preview_pen: u8,
-	/// Active pen preview.
-	pub preview_pen: RsnapFrozenOverlayExportElement,
-	/// Non-zero when `preview_arrow` is present.
-	pub has_preview_arrow: u8,
-	/// Active arrow preview.
-	pub preview_arrow: RsnapFrozenOverlayExportElement,
-	/// Non-zero when `preview_mosaic` is present.
-	pub has_preview_mosaic: u8,
-	/// Active mosaic preview.
-	pub preview_mosaic: RsnapFrozenOverlayExportElement,
-	/// Non-zero when `preview_spotlight` is present.
-	pub has_preview_spotlight: u8,
-	/// Active spotlight preview.
-	pub preview_spotlight: RsnapFrozenOverlayExportElement,
-	/// Non-zero when `preview_text` is present.
-	pub has_preview_text: u8,
-	/// Active moved text preview.
-	pub preview_text: RsnapFrozenOverlayExportElement,
-	/// Non-zero when `active_text_edit` is present.
-	pub has_active_text_edit: u8,
-	/// Active text edit payload.
-	pub active_text_edit: RsnapFrozenOverlayExportElement,
-}
-impl Default for RsnapFrozenOverlayEditSnapshot {
-	fn default() -> Self {
-		Self {
-			can_undo: 0,
-			can_redo: 0,
-			keeps_frozen_selection_fixed: 0,
-			is_moving_movable_annotation: 0,
-			has_active_interaction: 0,
-			elements: ptr::null_mut(),
-			elements_len: 0,
-			has_preview_pen: 0,
-			preview_pen: frozen_overlay_empty_element(),
-			has_preview_arrow: 0,
-			preview_arrow: frozen_overlay_empty_element(),
-			has_preview_mosaic: 0,
-			preview_mosaic: frozen_overlay_empty_element(),
-			has_preview_spotlight: 0,
-			preview_spotlight: frozen_overlay_empty_element(),
-			has_preview_text: 0,
-			preview_text: frozen_overlay_empty_element(),
-			has_active_text_edit: 0,
-			active_text_edit: frozen_overlay_empty_element(),
-		}
-	}
-}
-
-/// FFI-safe capture-frame source discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapCaptureFrameSourceKind {
-	/// User-dragged region capture.
-	DragRegion = 0,
-	/// Single-window capture.
-	Window = 1,
-	/// Full-screen capture.
-	FullScreen = 2,
-	/// Scroll-capture export.
-	ScrollCapture = 3,
-	/// Unknown or future capture source.
-	Unknown = 4,
-}
-
-/// FFI-safe capture-frame background discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapCaptureFrameBackgroundKind {
-	/// Prefer system wallpaper with gradient fallback.
-	SystemWallpaper = 0,
-	/// Blue-to-warm product gradient.
-	Aurora = 1,
-	/// Neutral graphite gradient.
-	Graphite = 2,
-	/// Light linen gradient.
-	Linen = 3,
-}
-
-/// FFI-safe capture-frame render mode.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapCaptureFrameRenderKind {
-	/// Draw shadows and rounded clipping around the capture.
-	FramedCapture = 0,
-	/// Draw a floating full-window snapshot without added clipping.
-	WindowSnapshot = 1,
-}
-
-/// FFI-safe sRGB capture-frame color stop.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapCaptureFrameColorStop {
-	/// Red component in sRGB space.
-	pub red: f64,
-	/// Green component in sRGB space.
-	pub green: f64,
-	/// Blue component in sRGB space.
-	pub blue: f64,
-	/// Alpha component.
-	pub alpha: f64,
-}
-
-/// FFI-safe capture-frame background drawing plan.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapCaptureFrameBackgroundPlan {
-	/// Ordered sRGB gradient color stops.
-	pub colors: [RsnapCaptureFrameColorStop; 3],
-	/// Gradient locations matching `colors`.
-	pub locations: [f64; 3],
-	/// Non-zero when the host should first try drawing system wallpaper.
-	pub prefers_wallpaper: u8,
-	/// Overlay alpha applied when wallpaper drawing succeeds.
-	pub wallpaper_overlay_alpha: f64,
-}
-
-/// FFI-safe capture-frame shadow pass.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapCaptureFrameShadow {
-	/// Horizontal shadow offset in output pixels.
-	pub offset_x: f64,
-	/// Vertical shadow offset in output pixels.
-	pub offset_y: f64,
-	/// Shadow blur radius in output pixels.
-	pub blur: f64,
-	/// Shadow alpha.
-	pub alpha: f64,
-}
-
-/// FFI-safe capture-frame drawing plan.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapCaptureFramePlan {
-	/// Canvas width in output pixels.
-	pub canvas_width: f64,
-	/// Canvas height in output pixels.
-	pub canvas_height: f64,
-	/// Image placement inside the canvas.
-	pub image_rect: RsnapFloatRect,
-	/// Rounded capture corner radius.
-	pub corner_radius: f64,
-	/// Ordered shadow passes behind the framed capture.
-	pub shadows: [RsnapCaptureFrameShadow; 3],
-}
-
-/// FFI-safe platform wallpaper thumbnail request.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapCaptureFrameWallpaperRequest {
-	/// Maximum thumbnail dimension requested from the platform image pipeline.
-	pub target_pixel_size: u32,
-	/// Overlay alpha applied after drawing the wallpaper thumbnail.
-	pub overlay_alpha: f64,
-}
-
-/// FFI-safe scroll-capture minimap layout plan.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct RsnapScrollMinimapPlan {
-	/// Outer minimap frame.
-	pub frame: RsnapFloatRect,
-	/// Preview image frame inside `frame`.
-	pub image_frame: RsnapFloatRect,
-	/// Non-zero when `viewport_frame` contains a visible marker.
-	pub has_viewport_frame: u8,
-	/// Viewport marker frame inside `image_frame`.
-	pub viewport_frame: RsnapFloatRect,
-}
-
-/// FFI-safe frozen selection transform discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapFrozenSelectionTransformKind {
-	/// Move the whole selection rectangle.
-	Move = 0,
-	/// Resize the left edge.
-	ResizeLeft = 1,
-	/// Resize the right edge.
-	ResizeRight = 2,
-	/// Resize the top edge.
-	ResizeTop = 3,
-	/// Resize the bottom edge.
-	ResizeBottom = 4,
-	/// Resize the top-left corner.
-	ResizeTopLeft = 5,
-	/// Resize the top-right corner.
-	ResizeTopRight = 6,
-	/// Resize the bottom-left corner.
-	ResizeBottomLeft = 7,
-	/// Resize the bottom-right corner.
-	ResizeBottomRight = 8,
-}
-
-/// FFI-safe scroll-capture observation discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapScrollObserveOutcomeKind {
-	/// The candidate did not change committed output.
-	NoChange = 0,
-	/// Preview-only state changed.
-	PreviewUpdated = 1,
-	/// Downward growth was committed.
-	Committed = 2,
-	/// The candidate proved motion in a direction not appended by this wrapper.
-	UnsupportedDirection = 3,
-}
-
-/// FFI-safe result for one scroll-capture frame observation.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapScrollObserveResult {
-	/// Observation outcome.
-	pub kind: u32,
-	/// Appended row count when `kind` is committed.
-	pub growth_rows: u32,
-	/// Current committed export width in pixels.
-	pub export_width: u32,
-	/// Current committed export height in pixels.
-	pub export_height: u32,
-	/// Current committed viewport top in pixels.
-	pub current_viewport_top_y: i32,
-}
-impl Default for RsnapScrollObserveResult {
-	fn default() -> Self {
-		Self {
-			kind: RsnapScrollObserveOutcomeKind::NoChange as u32,
-			growth_rows: 0,
-			export_width: 0,
-			export_height: 0,
-			current_viewport_top_y: 0,
-		}
-	}
-}
-
-/// FFI-safe platform tag.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapPlatformTag {
-	/// Native macOS host.
-	MacOS = 0,
-	/// Future native Windows host.
-	Windows = 1,
-	/// Future native Linux host.
-	Linux = 2,
-	/// Unsupported or test-only host.
-	Unsupported = 3,
-}
-
-/// FFI-safe session configuration.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapSessionConfig {
-	/// Platform family that owns the host.
-	pub platform: RsnapPlatformTag,
-	/// Non-zero when native text input is available.
-	pub allow_text_input: u8,
-	/// Non-zero when the host prefers the toolbar above the frozen selection.
-	pub prefers_toolbar_above_selection: u8,
-}
-
-/// FFI-safe global point.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapPoint {
-	/// Global X coordinate.
-	pub x: i32,
-	/// Global Y coordinate.
-	pub y: i32,
-}
-
-/// FFI-safe RGB sample.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapRgb {
-	/// Red channel.
-	pub r: u8,
-	/// Green channel.
-	pub g: u8,
-	/// Blue channel.
-	pub b: u8,
-}
-
-/// FFI-safe global rectangle.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapRect {
-	/// Global left coordinate.
-	pub x: i32,
-	/// Global top coordinate.
-	pub y: i32,
-	/// Rectangle width.
-	pub width: u32,
-	/// Rectangle height.
-	pub height: u32,
-}
-
-/// FFI-safe monitor rectangle snapshot.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapMonitorRect {
-	/// Stable monitor identifier.
-	pub id: u32,
-	/// Monitor origin in global points.
-	pub origin: RsnapPoint,
-	/// Monitor width in points.
-	pub width: u32,
-	/// Monitor height in points.
-	pub height: u32,
-	/// Monitor pixel scale factor in thousandths.
-	pub scale_factor_x1000: u32,
-}
-
-/// FFI-safe highlighted window snapshot.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapWindowRect {
-	/// Window identifier when one exists.
-	pub window_id: u32,
-	/// Non-zero when `window_id` is present.
-	pub has_window_id: u8,
-	/// Global left coordinate.
-	pub x: i64,
-	/// Global top coordinate.
-	pub y: i64,
-	/// Window width in points.
-	pub width: i64,
-	/// Window height in points.
-	pub height: i64,
-}
-
-/// FFI-safe host event discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapHostEventKind {
-	/// Capture session became active.
-	SessionActivated = 0,
-	/// Pointer position changed.
-	PointerMoved = 1,
-	/// Cancel was requested.
-	CancelRequested = 3,
-	/// Copy was requested.
-	CopyRequested = 4,
-	/// Save was requested.
-	SaveRequested = 5,
-	/// Text recognition was requested.
-	RecognizeTextRequested = 6,
-	/// Loupe visibility was toggled.
-	ToggleLoupe = 7,
-	/// A frozen toolbar item was invoked.
-	ToolbarItemInvoked = 8,
-	/// A primary interaction began in live mode.
-	PrimaryInteractionStarted = 9,
-	/// A primary interaction updated in live mode.
-	PrimaryInteractionUpdated = 10,
-	/// A primary interaction completed in live mode.
-	PrimaryInteractionCompleted = 11,
-}
-
-/// FFI-safe frozen toolbar item kind.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapToolbarItemKind {
-	/// Pointer tool.
-	Pointer = 0,
-	/// Pen tool.
-	Pen = 1,
-	/// Arrow tool.
-	Arrow = 2,
-	/// Text tool.
-	Text = 3,
-	/// Mosaic tool.
-	Mosaic = 4,
-	/// Spotlight tool.
-	Spotlight = 5,
-	/// Undo action.
-	Undo = 6,
-	/// Redo action.
-	Redo = 7,
-	/// Auto-center action.
-	AutoCenter = 8,
-	/// Scroll capture action.
-	Scroll = 9,
-	/// OCR action.
-	Ocr = 10,
-	/// Copy action.
-	Copy = 11,
-	/// Save action.
-	Save = 12,
-}
-
-/// FFI-safe host event payload.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapHostEvent {
-	/// Event discriminator.
-	pub kind: u32,
-	/// Pointer location when the event carries one.
-	pub point: RsnapPoint,
-	/// Non-zero when `point` is present.
-	pub has_point: u8,
-	/// RGB sample when the event carries one.
-	pub rgb: RsnapRgb,
-	/// Non-zero when `rgb` is present.
-	pub has_rgb: u8,
-	/// Current active monitor when the event carries one.
-	pub active_monitor: RsnapMonitorRect,
-	/// Non-zero when `active_monitor` is present.
-	pub has_active_monitor: u8,
-	/// Highlighted live window when the event carries one.
-	pub highlighted_window: RsnapWindowRect,
-	/// Non-zero when `highlighted_window` is present.
-	pub has_highlighted_window: u8,
-	/// Toolbar item kind when the event carries one.
-	pub toolbar_item_kind: u32,
-}
-
-/// FFI-safe host report discriminator.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapHostReportKind {
-	/// Frozen snapshot committed.
-	FreezeSnapshotCommitted = 0,
-	/// Host effect completed.
-	HostEffectCompleted = 1,
-	/// Permission state changed.
-	PermissionChanged = 2,
-	/// Host surfaced a status message.
-	StatusMessage = 3,
-}
-
-/// FFI-safe host effect kind.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapHostEffectKind {
-	/// Copy capture to the clipboard.
-	CopyCapture = 0,
-	/// Save capture to disk.
-	SaveCapture = 1,
-	/// Recognize text from the capture.
-	RecognizeText = 2,
-}
-
-/// FFI-safe permission kind.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapPermissionKind {
-	/// Screen recording or equivalent display capture access.
-	ScreenRecording = 0,
-}
-
-/// FFI-safe host report payload.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapHostReport {
-	/// Report discriminator.
-	pub kind: u32,
-	/// Frozen selection rectangle when the report carries one.
-	pub selection: RsnapRect,
-	/// Non-zero when `selection` is present.
-	pub has_selection: u8,
-	/// Host effect kind when the report carries one.
-	pub effect_kind: u32,
-	/// Permission kind when the report carries one.
-	pub permission_kind: u32,
-	/// Non-zero when the permission is granted.
-	pub granted: u8,
-	/// UTF-8 status message byte count copied into `status_message`.
-	pub status_message_len: u32,
-	/// Optional UTF-8 status message bytes.
-	pub status_message: [u8; RSNAP_STATUS_MESSAGE_CAPACITY],
-}
-
-/// FFI-safe scene kind.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapSceneKind {
-	/// Capture UI is hidden.
-	Hidden = 0,
-	/// Live targeting is active.
-	Live = 1,
-	/// Frozen mode is active.
-	Frozen = 2,
-}
-
-/// FFI-safe cursor intent.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapCursorIntent {
-	/// Default cursor.
-	Default = 0,
-	/// Crosshair cursor.
-	Crosshair = 1,
-	/// Grab cursor.
-	Grab = 2,
-	/// Grabbing cursor.
-	Grabbing = 3,
-	/// North resize cursor.
-	ResizeNorth = 4,
-	/// South resize cursor.
-	ResizeSouth = 5,
-	/// East resize cursor.
-	ResizeEast = 6,
-	/// West resize cursor.
-	ResizeWest = 7,
-	/// North-east resize cursor.
-	ResizeNorthEast = 8,
-	/// North-west resize cursor.
-	ResizeNorthWest = 9,
-	/// South-east resize cursor.
-	ResizeSouthEast = 10,
-	/// South-west resize cursor.
-	ResizeSouthWest = 11,
-	/// Text cursor.
-	Text = 12,
-}
-
-/// FFI-safe scene snapshot copy.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct RsnapSceneModel {
-	/// Scene kind.
-	pub scene_kind: u32,
-	/// Cursor intent.
-	pub cursor_intent: u32,
-	/// Current pointer position.
-	pub pointer: RsnapPoint,
-	/// Non-zero when `pointer` is present.
-	pub has_pointer: u8,
-	/// Current active monitor.
-	pub active_monitor: RsnapMonitorRect,
-	/// Non-zero when `active_monitor` is present.
-	pub has_active_monitor: u8,
-	/// Highlighted live window.
-	pub highlighted_window: RsnapWindowRect,
-	/// Non-zero when `highlighted_window` is present.
-	pub has_highlighted_window: u8,
-	/// Live drag preview rectangle.
-	pub live_selection_preview: RsnapRect,
-	/// Non-zero when `live_selection_preview` is present.
-	pub has_live_selection_preview: u8,
-	/// Frozen selection rectangle.
-	pub frozen_selection: RsnapRect,
-	/// Non-zero when `frozen_selection` is present.
-	pub has_frozen_selection: u8,
-	/// Current RGB sample.
-	pub rgb: RsnapRgb,
-	/// Non-zero when `rgb` is present.
-	pub has_rgb: u8,
-	/// Non-zero when the loupe should be visible.
-	pub loupe_visible: u8,
-	/// Toolbar item count copied into `toolbar_items`.
-	pub toolbar_item_count: u32,
-	/// Frozen toolbar items.
-	pub toolbar_items: [RsnapToolbarItem; RSNAP_TOOLBAR_ITEM_CAPACITY],
-	/// UTF-8 status message byte count copied into `status_message`.
-	pub status_message_len: u32,
-	/// Optional UTF-8 status message bytes.
-	pub status_message: [u8; RSNAP_STATUS_MESSAGE_CAPACITY],
-}
-impl Default for RsnapSceneModel {
-	fn default() -> Self {
-		Self {
-			scene_kind: 0,
-			cursor_intent: 0,
-			pointer: RsnapPoint::default(),
-			has_pointer: 0,
-			active_monitor: RsnapMonitorRect::default(),
-			has_active_monitor: 0,
-			highlighted_window: RsnapWindowRect::default(),
-			has_highlighted_window: 0,
-			live_selection_preview: RsnapRect::default(),
-			has_live_selection_preview: 0,
-			frozen_selection: RsnapRect::default(),
-			has_frozen_selection: 0,
-			rgb: RsnapRgb::default(),
-			has_rgb: 0,
-			loupe_visible: 0,
-			toolbar_item_count: 0,
-			toolbar_items: [RsnapToolbarItem::default(); RSNAP_TOOLBAR_ITEM_CAPACITY],
-			status_message_len: 0,
-			status_message: [0; RSNAP_STATUS_MESSAGE_CAPACITY],
-		}
-	}
-}
-
-/// One FFI-safe frozen toolbar item.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapToolbarItem {
-	/// Item kind.
-	pub kind: u32,
-	/// Non-zero when the item is enabled.
-	pub enabled: u8,
-	/// Non-zero when the item is selected.
-	pub selected: u8,
-	/// Non-zero when the slot is populated.
-	pub present: u8,
-}
-
-/// FFI-safe host request kind.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum RsnapHostRequestKind {
-	/// Start native live capture.
-	StartLiveCapture = 0,
-	/// Stop native live capture.
-	StopLiveCapture = 1,
-	/// Request a frozen snapshot.
-	RequestFreezeSnapshot = 2,
-	/// Copy capture effect.
-	CopyCapture = 3,
-	/// Save capture effect.
-	SaveCapture = 4,
-	/// Recognize text effect.
-	RecognizeText = 5,
-	/// Request screen recording permission.
-	RequestScreenRecordingPermission = 6,
-	/// Start native scroll capture.
-	StartScrollCapture = 9,
-}
-
-/// FFI-safe queued host request.
-#[repr(C)]
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-pub struct RsnapHostRequestValue {
-	/// Request kind.
-	pub kind: u32,
-	/// Optional selection payload for frozen handoff requests.
-	pub selection: RsnapRect,
-	/// Non-zero when `selection` is populated.
-	pub has_selection: u8,
-	/// Non-zero when the frozen selection may be moved or resized after commit.
-	pub selection_editable: u8,
-}
 
 /// Creates a new opaque session handle.
 ///
@@ -3229,22 +2321,6 @@ fn decode_frozen_annotation_color(color: RsnapFrozenAnnotationColor) -> [u8; 4] 
 	}
 }
 
-fn frozen_overlay_empty_element() -> RsnapFrozenOverlayExportElement {
-	RsnapFrozenOverlayExportElement {
-		kind: RsnapFrozenOverlayExportElementKind::Mosaic,
-		rect: RsnapFloatRect::default(),
-		start: RsnapFloatPoint::default(),
-		end: RsnapFloatPoint::default(),
-		points: ptr::null(),
-		points_len: 0,
-		text: ptr::null(),
-		stroke_width_points: 0.0,
-		border_width_points: 0.0,
-		font_size_points: 0.0,
-		color: RsnapFrozenAnnotationColor::Blue,
-	}
-}
-
 fn decode_frozen_overlay_edit_style(
 	style: RsnapFrozenOverlayEditStyle,
 ) -> Option<FrozenOverlayEditStyle> {
@@ -3338,17 +2414,17 @@ fn encode_frozen_overlay_edit_snapshot(
 		preview_arrow: snapshot
 			.preview_arrow
 			.map(encode_frozen_overlay_edit_arrow)
-			.unwrap_or_else(frozen_overlay_empty_element),
+			.unwrap_or_else(abi::frozen_overlay_empty_element),
 		has_preview_mosaic: u8::from(snapshot.preview_mosaic.is_some()),
 		preview_mosaic: snapshot
 			.preview_mosaic
 			.map(encode_frozen_overlay_edit_mosaic)
-			.unwrap_or_else(frozen_overlay_empty_element),
+			.unwrap_or_else(abi::frozen_overlay_empty_element),
 		has_preview_spotlight: u8::from(snapshot.preview_spotlight.is_some()),
 		preview_spotlight: snapshot
 			.preview_spotlight
 			.map(encode_frozen_overlay_edit_spotlight)
-			.unwrap_or_else(frozen_overlay_empty_element),
+			.unwrap_or_else(abi::frozen_overlay_empty_element),
 		has_preview_text: u8::from(snapshot.preview_text.is_some()),
 		preview_text: encode_optional_frozen_overlay_edit_text(snapshot.preview_text.as_ref())?,
 		has_active_text_edit: u8::from(snapshot.active_text_edit.is_some()),
@@ -3387,14 +2463,15 @@ fn encode_frozen_overlay_edit_pen(
 		points_len,
 		stroke_width_points: annotation.style.stroke_width_points,
 		color: encode_frozen_edit_color(annotation.style.color),
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	})
 }
 
 fn encode_optional_frozen_overlay_edit_pen(
 	annotation: Option<&FrozenOverlayEditPen>,
 ) -> Option<RsnapFrozenOverlayExportElement> {
-	annotation.map_or_else(|| Some(frozen_overlay_empty_element()), encode_frozen_overlay_edit_pen)
+	annotation
+		.map_or_else(|| Some(abi::frozen_overlay_empty_element()), encode_frozen_overlay_edit_pen)
 }
 
 fn encode_frozen_overlay_edit_arrow(
@@ -3406,7 +2483,7 @@ fn encode_frozen_overlay_edit_arrow(
 		end: encode_frozen_edit_point(annotation.end),
 		stroke_width_points: annotation.style.stroke_width_points,
 		color: encode_frozen_edit_color(annotation.style.color),
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	}
 }
 
@@ -3416,7 +2493,7 @@ fn encode_frozen_overlay_edit_mosaic(
 	RsnapFrozenOverlayExportElement {
 		kind: RsnapFrozenOverlayExportElementKind::Mosaic,
 		rect: encode_frozen_edit_rect(annotation.rect),
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	}
 }
 
@@ -3428,7 +2505,7 @@ fn encode_frozen_overlay_edit_spotlight(
 		rect: encode_frozen_edit_rect(annotation.rect),
 		border_width_points: annotation.style.border_width_points,
 		color: encode_frozen_edit_color(annotation.style.border_color),
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	}
 }
 
@@ -3443,14 +2520,15 @@ fn encode_frozen_overlay_edit_text(
 		text,
 		font_size_points: annotation.style.font_size_points,
 		color: encode_frozen_edit_color(annotation.style.color),
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	})
 }
 
 fn encode_optional_frozen_overlay_edit_text(
 	annotation: Option<&FrozenOverlayEditText>,
 ) -> Option<RsnapFrozenOverlayExportElement> {
-	annotation.map_or_else(|| Some(frozen_overlay_empty_element()), encode_frozen_overlay_edit_text)
+	annotation
+		.map_or_else(|| Some(abi::frozen_overlay_empty_element()), encode_frozen_overlay_edit_text)
 }
 
 fn encode_frozen_overlay_active_text_edit(
@@ -3462,7 +2540,7 @@ fn encode_frozen_overlay_active_text_edit(
 		kind: RsnapFrozenOverlayExportElementKind::Text,
 		start: encode_frozen_edit_point(edit.anchor),
 		text,
-		..frozen_overlay_empty_element()
+		..abi::frozen_overlay_empty_element()
 	})
 }
 
@@ -3470,7 +2548,7 @@ fn encode_optional_frozen_overlay_active_text_edit(
 	edit: Option<&FrozenOverlayTextEdit>,
 ) -> Option<RsnapFrozenOverlayExportElement> {
 	edit.map_or_else(
-		|| Some(frozen_overlay_empty_element()),
+		|| Some(abi::frozen_overlay_empty_element()),
 		encode_frozen_overlay_active_text_edit,
 	)
 }
@@ -3534,7 +2612,7 @@ unsafe fn release_frozen_overlay_snapshot_element(element: &mut RsnapFrozenOverl
 		let _ = unsafe { CString::from_raw(element.text as *mut c_char) };
 	}
 
-	*element = frozen_overlay_empty_element();
+	*element = abi::frozen_overlay_empty_element();
 }
 
 unsafe fn decode_bgra_frame<'a>(
@@ -4182,1297 +3260,4 @@ fn encode_window(window: WindowRect) -> RsnapWindowRect {
 }
 
 #[cfg(test)]
-mod tests {
-	use std::env;
-	use std::ffi::CString;
-	use std::fs;
-	use std::process;
-	use std::ptr;
-	use std::slice;
-
-	use crate::{
-		RSNAP_HOST_FFI_ABI_VERSION, RSNAP_STATUS_MESSAGE_CAPACITY, RsnapCaptureFrameBackgroundKind,
-		RsnapCaptureFrameBackgroundPlan, RsnapCaptureFrameColorStop, RsnapCaptureFramePlan,
-		RsnapCaptureFrameRenderKind, RsnapCaptureFrameSourceKind,
-		RsnapCaptureFrameWallpaperRequest, RsnapCursorIntent, RsnapFloatPoint, RsnapFloatRect,
-		RsnapFrozenAnnotationColor, RsnapFrozenOverlayEditSnapshot, RsnapFrozenOverlayEditStyle,
-		RsnapFrozenOverlayExportElement, RsnapFrozenOverlayExportElementKind,
-		RsnapFrozenSelectionTransformKind, RsnapHostEvent, RsnapHostEventKind, RsnapHostReport,
-		RsnapHostReportKind, RsnapHostRequestKind, RsnapHostRequestValue, RsnapMonitorRect,
-		RsnapOwnedBytes, RsnapOwnedRgbaRegion, RsnapPixelRect, RsnapPlatformTag, RsnapPoint,
-		RsnapRect, RsnapRgb, RsnapSceneKind, RsnapSceneModel, RsnapScrollMinimapPlan,
-		RsnapSessionConfig, RsnapSessionHandle, RsnapStatus, RsnapToolbarItemKind, RsnapWindowRect,
-	};
-	#[cfg(target_os = "macos")]
-	use crate::{RsnapScrollObserveOutcomeKind, RsnapScrollObserveResult};
-
-	fn default_config() -> RsnapSessionConfig {
-		RsnapSessionConfig {
-			platform: RsnapPlatformTag::MacOS,
-			allow_text_input: 1,
-			prefers_toolbar_above_selection: 0,
-		}
-	}
-
-	fn scroll_frame(width: u32, height: u32, top_row: u32) -> Vec<u8> {
-		let mut rgba = Vec::with_capacity((width * height * 4) as usize);
-
-		for y in 0..height {
-			let document_row = top_row + y;
-
-			for x in 0..width {
-				rgba.push(((document_row * 17 + x * 13) % 251) as u8);
-				rgba.push(((document_row * 29 + x * 7) % 251) as u8);
-				rgba.push(((document_row * 5 + x * 31) % 251) as u8);
-				rgba.push(255);
-			}
-		}
-
-		rgba
-	}
-
-	fn png_dimensions(png: &[u8]) -> (u32, u32) {
-		assert!(png.starts_with(b"\x89PNG\r\n\x1a\n"));
-
-		let width = u32::from_be_bytes(png[16..20].try_into().expect("PNG width bytes"));
-		let height = u32::from_be_bytes(png[20..24].try_into().expect("PNG height bytes"));
-
-		(width, height)
-	}
-
-	fn png_phys_chunk(png: &[u8]) -> Option<(u32, u32, u8)> {
-		assert!(png.starts_with(b"\x89PNG\r\n\x1a\n"));
-
-		let mut offset = 8;
-
-		while offset + 12 <= png.len() {
-			let length = u32::from_be_bytes(
-				png[offset..offset + 4].try_into().expect("PNG chunk length bytes"),
-			) as usize;
-			let data_start = offset + 8;
-			let data_end = data_start.checked_add(length)?;
-			let next_offset = data_end.checked_add(4)?;
-
-			if next_offset > png.len() {
-				return None;
-			}
-			if &png[offset + 4..offset + 8] == b"pHYs" {
-				assert_eq!(length, 9);
-
-				let x = u32::from_be_bytes(
-					png[data_start..data_start + 4].try_into().expect("pHYs x bytes"),
-				);
-				let y = u32::from_be_bytes(
-					png[data_start + 4..data_start + 8].try_into().expect("pHYs y bytes"),
-				);
-
-				return Some((x, y, png[data_start + 8]));
-			}
-
-			offset = next_offset;
-		}
-
-		None
-	}
-
-	#[test]
-	fn ffi_session_enters_live_and_emits_request() {
-		let handle = unsafe { crate::rsnap_session_create(default_config()) };
-		let mut request =
-			RsnapHostRequestValue { kind: u32::MAX, ..RsnapHostRequestValue::default() };
-		let mut scene = RsnapSceneModel::default();
-
-		assert_eq!(unsafe { crate::rsnap_session_enter_live(handle) }, RsnapStatus::Ok);
-		assert_eq!(
-			unsafe { crate::rsnap_session_take_next_request(handle, &mut request) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(request.kind, RsnapHostRequestKind::StartLiveCapture as u32);
-		assert_eq!(
-			unsafe { crate::rsnap_session_copy_scene_model(handle, &mut scene) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(scene.scene_kind, RsnapSceneKind::Live as u32);
-		assert_eq!(scene.cursor_intent, RsnapCursorIntent::Default as u32);
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn ffi_session_applies_freeze_report() {
-		let handle = unsafe { crate::rsnap_session_create(default_config()) };
-		let mut scene = RsnapSceneModel::default();
-
-		assert_eq!(unsafe { crate::rsnap_session_enter_live(handle) }, RsnapStatus::Ok);
-
-		let mut request =
-			RsnapHostRequestValue { kind: u32::MAX, ..RsnapHostRequestValue::default() };
-		let _ = unsafe { crate::rsnap_session_take_next_request(handle, &mut request) };
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_session_handle_host_report(
-					handle,
-					RsnapHostReport {
-						kind: RsnapHostReportKind::FreezeSnapshotCommitted as u32,
-						selection: RsnapRect { x: 20, y: 30, width: 100, height: 60 },
-						has_selection: 1,
-						effect_kind: 0,
-						permission_kind: 0,
-						granted: 0,
-						status_message_len: 0,
-						status_message: [0; RSNAP_STATUS_MESSAGE_CAPACITY],
-					},
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(
-			unsafe { crate::rsnap_session_copy_scene_model(handle, &mut scene) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(scene.scene_kind, RsnapSceneKind::Frozen as u32);
-		assert_eq!(scene.has_frozen_selection, 1);
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn ffi_session_tracks_pointer_updates() {
-		let handle = unsafe { crate::rsnap_session_create(default_config()) };
-		let mut scene = RsnapSceneModel::default();
-		let _ = unsafe { crate::rsnap_session_enter_live(handle) };
-		let mut request =
-			RsnapHostRequestValue { kind: u32::MAX, ..RsnapHostRequestValue::default() };
-		let _ = unsafe { crate::rsnap_session_take_next_request(handle, &mut request) };
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_session_handle_host_event(
-					handle,
-					RsnapHostEvent {
-						kind: RsnapHostEventKind::PointerMoved as u32,
-						point: RsnapPoint { x: 50, y: 60 },
-						has_point: 1,
-						rgb: RsnapRgb { r: 1, g: 2, b: 3 },
-						has_rgb: 1,
-						active_monitor: RsnapMonitorRect {
-							id: 9,
-							origin: RsnapPoint { x: 0, y: 0 },
-							width: 1_440,
-							height: 900,
-							scale_factor_x1000: 2_000,
-						},
-						has_active_monitor: 1,
-						highlighted_window: RsnapWindowRect {
-							window_id: 42,
-							has_window_id: 1,
-							x: 20,
-							y: 30,
-							width: 500,
-							height: 400,
-						},
-						has_highlighted_window: 1,
-						toolbar_item_kind: 0,
-					},
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(
-			unsafe { crate::rsnap_session_copy_scene_model(handle, &mut scene) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(scene.has_pointer, 1);
-		assert_eq!(scene.pointer.x, 50);
-		assert_eq!(scene.has_rgb, 1);
-		assert_eq!(scene.has_active_monitor, 1);
-		assert_eq!(scene.active_monitor.id, 9);
-		assert_eq!(scene.has_highlighted_window, 1);
-		assert_eq!(scene.highlighted_window.window_id, 42);
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn destroy_allows_null() {
-		let handle: *mut RsnapSessionHandle = ptr::null_mut();
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn ffi_export_rgba_to_png_returns_owned_png() {
-		let rgba = scroll_frame(4, 4, 0);
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe { crate::rsnap_export_rgba_to_png(4, 4, rgba.as_ptr(), rgba.len(), &mut png) },
-			RsnapStatus::Ok
-		);
-		assert!(png.len > 0);
-
-		let png_bytes = unsafe { slice::from_raw_parts(png.bytes, png.len) };
-
-		assert_eq!(png_dimensions(png_bytes), (4, 4));
-		assert_eq!(png_phys_chunk(png_bytes), None);
-
-		unsafe {
-			crate::rsnap_owned_bytes_release(&mut png);
-		}
-
-		assert!(png.bytes.is_null());
-		assert_eq!(png.len, 0);
-		assert_eq!(png.capacity, 0);
-	}
-
-	#[test]
-	fn ffi_export_rgba_to_png_with_screen_scale_writes_density() {
-		let rgba = scroll_frame(4, 4, 0);
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_export_rgba_to_png_with_screen_scale(
-					4,
-					4,
-					rgba.as_ptr(),
-					rgba.len(),
-					2_000,
-					&mut png,
-				)
-			},
-			RsnapStatus::Ok
-		);
-
-		let png_bytes = unsafe { slice::from_raw_parts(png.bytes, png.len) };
-
-		assert_eq!(png_dimensions(png_bytes), (4, 4));
-		assert_eq!(png_phys_chunk(png_bytes), Some((5_669, 5_669, 1)));
-
-		unsafe {
-			crate::rsnap_owned_bytes_release(&mut png);
-		}
-	}
-
-	#[test]
-	fn ffi_export_rgba_to_png_with_screen_scale_rejects_zero_scale() {
-		let rgba = scroll_frame(4, 4, 0);
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_export_rgba_to_png_with_screen_scale(
-					4,
-					4,
-					rgba.as_ptr(),
-					rgba.len(),
-					0,
-					&mut png,
-				)
-			},
-			RsnapStatus::InvalidInput
-		);
-		assert!(png.bytes.is_null());
-	}
-
-	#[test]
-	fn ffi_export_rgba_crop_to_png_crops_dimensions() {
-		let rgba = scroll_frame(4, 4, 0);
-		let crop = RsnapPixelRect { x: 1, y: 0, width: 2, height: 3 };
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_export_rgba_crop_to_png(
-					4,
-					4,
-					rgba.as_ptr(),
-					rgba.len(),
-					crop,
-					&mut png,
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(png_dimensions(unsafe { slice::from_raw_parts(png.bytes, png.len) }), (2, 3));
-
-		unsafe {
-			crate::rsnap_owned_bytes_release(&mut png);
-		}
-	}
-
-	#[test]
-	fn ffi_export_rgba_crop_to_png_with_screen_scale_writes_density() {
-		let rgba = scroll_frame(4, 4, 0);
-		let crop = RsnapPixelRect { x: 1, y: 0, width: 2, height: 3 };
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_export_rgba_crop_to_png_with_screen_scale(
-					4,
-					4,
-					rgba.as_ptr(),
-					rgba.len(),
-					crop,
-					2_000,
-					&mut png,
-				)
-			},
-			RsnapStatus::Ok
-		);
-
-		let png_bytes = unsafe { slice::from_raw_parts(png.bytes, png.len) };
-
-		assert_eq!(png_dimensions(png_bytes), (2, 3));
-		assert_eq!(png_phys_chunk(png_bytes), Some((5_669, 5_669, 1)));
-
-		unsafe {
-			crate::rsnap_owned_bytes_release(&mut png);
-		}
-	}
-
-	#[test]
-	fn ffi_export_rgba_crop_to_png_rejects_out_of_bounds_crop() {
-		let rgba = scroll_frame(4, 4, 0);
-		let crop = RsnapPixelRect { x: 3, y: 3, width: 2, height: 2 };
-		let mut png = RsnapOwnedBytes::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_export_rgba_crop_to_png(
-					4,
-					4,
-					rgba.as_ptr(),
-					rgba.len(),
-					crop,
-					&mut png,
-				)
-			},
-			RsnapStatus::InvalidInput
-		);
-		assert!(png.bytes.is_null());
-	}
-
-	#[test]
-	fn ffi_frozen_display_crop_rect_returns_core_pixel_rect() {
-		let mut out_rect = RsnapPixelRect::default();
-		let status = unsafe {
-			crate::rsnap_frozen_display_crop_rect(
-				2_880,
-				1_800,
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 1_440.0, height: 900.0 },
-				RsnapFloatRect { x: 100.0, y: 200.0, width: 300.0, height: 150.0 },
-				&mut out_rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(out_rect, RsnapPixelRect { x: 200, y: 1_100, width: 600, height: 300 });
-	}
-
-	#[test]
-	fn ffi_frozen_display_crop_rect_returns_empty_for_outside_selection() {
-		let mut out_rect = RsnapPixelRect::default();
-		let status = unsafe {
-			crate::rsnap_frozen_display_crop_rect(
-				200,
-				200,
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 100.0, height: 100.0 },
-				RsnapFloatRect { x: 120.0, y: 10.0, width: 10.0, height: 20.0 },
-				&mut out_rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Empty);
-	}
-
-	#[test]
-	fn ffi_frozen_mosaic_light_privacy_patch_returns_rgba_region() {
-		let mut patch = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_frozen_mosaic_light_privacy_patch_rgba(
-				100,
-				80,
-				RsnapFloatRect { x: 4.2, y: 9.1, width: 28.4, height: 21.0 },
-				&mut patch,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(patch.width, 3);
-		assert_eq!(patch.height, 3);
-		assert_eq!(patch.len, 36);
-
-		let bytes = unsafe { slice::from_raw_parts(patch.rgba, patch.len) };
-
-		assert_eq!(&bytes[..12], &[211, 211, 211, 255, 205, 205, 205, 255, 202, 201, 199, 255]);
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut patch);
-		}
-	}
-
-	#[test]
-	fn ffi_frozen_mosaic_light_privacy_patch_returns_empty_for_outside_rect() {
-		let mut patch = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_frozen_mosaic_light_privacy_patch_rgba(
-				100,
-				80,
-				RsnapFloatRect { x: 120.0, y: 10.0, width: 10.0, height: 20.0 },
-				&mut patch,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Empty);
-	}
-
-	#[test]
-	fn ffi_frozen_overlay_export_render_rgba_returns_composited_region() {
-		let mut rgba = vec![180_u8; 64 * 40 * 4];
-
-		for alpha in (3..rgba.len()).step_by(4) {
-			rgba[alpha] = 255;
-		}
-
-		let points = [RsnapFloatPoint { x: 2.0, y: 2.0 }, RsnapFloatPoint { x: 24.0, y: 18.0 }];
-		let text = CString::new("Hi").expect("text has no interior nul");
-		let elements = [
-			RsnapFrozenOverlayExportElement {
-				kind: RsnapFrozenOverlayExportElementKind::Mosaic,
-				rect: RsnapFloatRect { x: 4.0, y: 4.0, width: 16.0, height: 10.0 },
-				start: RsnapFloatPoint::default(),
-				end: RsnapFloatPoint::default(),
-				points: ptr::null(),
-				points_len: 0,
-				text: ptr::null(),
-				stroke_width_points: 0.0,
-				border_width_points: 0.0,
-				font_size_points: 0.0,
-				color: RsnapFrozenAnnotationColor::Blue,
-			},
-			RsnapFrozenOverlayExportElement {
-				kind: RsnapFrozenOverlayExportElementKind::Pen,
-				rect: RsnapFloatRect::default(),
-				start: RsnapFloatPoint::default(),
-				end: RsnapFloatPoint::default(),
-				points: points.as_ptr(),
-				points_len: points.len(),
-				text: ptr::null(),
-				stroke_width_points: 2.0,
-				border_width_points: 0.0,
-				font_size_points: 0.0,
-				color: RsnapFrozenAnnotationColor::Blue,
-			},
-			RsnapFrozenOverlayExportElement {
-				kind: RsnapFrozenOverlayExportElementKind::Text,
-				rect: RsnapFloatRect::default(),
-				start: RsnapFloatPoint { x: 6.0, y: 24.0 },
-				end: RsnapFloatPoint::default(),
-				points: ptr::null(),
-				points_len: 0,
-				text: text.as_ptr(),
-				stroke_width_points: 0.0,
-				border_width_points: 0.0,
-				font_size_points: 12.0,
-				color: RsnapFrozenAnnotationColor::White,
-			},
-		];
-		let mut out = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_frozen_overlay_export_render_rgba(
-				64,
-				40,
-				rgba.as_ptr(),
-				rgba.len(),
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 64.0, height: 40.0 },
-				elements.as_ptr(),
-				elements.len(),
-				&mut out,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(out.width, 64);
-		assert_eq!(out.height, 40);
-		assert_eq!(out.len, rgba.len());
-
-		let bytes = unsafe { slice::from_raw_parts(out.rgba, out.len) };
-
-		assert_ne!(bytes, rgba.as_slice());
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut out);
-		}
-	}
-
-	#[test]
-	fn ffi_frozen_overlay_edit_session_copies_rust_owned_snapshot() {
-		let handle = crate::rsnap_frozen_overlay_edit_session_create();
-
-		assert!(!handle.is_null());
-
-		let style = RsnapFrozenOverlayEditStyle {
-			stroke_width_points: 3.0,
-			stroke_color: RsnapFrozenAnnotationColor::Blue,
-			spotlight_border_width_points: 0.0,
-			spotlight_color: RsnapFrozenAnnotationColor::Blue,
-			text_font_size_points: 16.0,
-			text_color: RsnapFrozenAnnotationColor::White,
-		};
-		let selection = RsnapFloatRect { x: 0.0, y: 0.0, width: 200.0, height: 120.0 };
-		let mut changed = 0;
-		let status = unsafe {
-			crate::rsnap_frozen_overlay_edit_session_begin(
-				handle,
-				RsnapToolbarItemKind::Text,
-				RsnapFloatPoint { x: 12.0, y: 18.0 },
-				selection,
-				style,
-				&mut changed,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(changed, 1);
-
-		let text = CString::new("Hello").expect("text has no interior nul");
-		let status = unsafe {
-			crate::rsnap_frozen_overlay_edit_session_append_text(
-				handle,
-				text.as_ptr(),
-				&mut changed,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(changed, 1);
-
-		let status = unsafe {
-			crate::rsnap_frozen_overlay_edit_session_commit_text(handle, style, &mut changed)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(changed, 1);
-
-		let mut snapshot = RsnapFrozenOverlayEditSnapshot::default();
-		let status = unsafe {
-			crate::rsnap_frozen_overlay_edit_session_copy_snapshot(handle, &mut snapshot)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(snapshot.can_undo, 1);
-		assert_eq!(snapshot.elements_len, 1);
-
-		let elements = unsafe { slice::from_raw_parts(snapshot.elements, snapshot.elements_len) };
-
-		assert_eq!(elements[0].kind, RsnapFrozenOverlayExportElementKind::Text);
-		assert_eq!(unsafe { std::ffi::CStr::from_ptr(elements[0].text) }.to_str(), Ok("Hello"));
-
-		unsafe {
-			crate::rsnap_frozen_overlay_edit_snapshot_release(&mut snapshot);
-			crate::rsnap_frozen_overlay_edit_session_destroy(handle);
-		}
-	}
-
-	#[test]
-	fn ffi_bgra_frame_sample_rgb_returns_core_sample() {
-		let bgra = bgra_frame(4, 3, 16);
-		let mut rgb = RsnapRgb::default();
-		let status = unsafe {
-			crate::rsnap_bgra_frame_sample_rgb(
-				4,
-				3,
-				16,
-				bgra.as_ptr(),
-				bgra.len(),
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 4.0, height: 3.0 },
-				1.0,
-				2.5,
-				&mut rgb,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(rgb, RsnapRgb { r: 11, g: 21, b: 31 });
-	}
-
-	#[test]
-	fn ffi_bgra_frame_loupe_patch_returns_rgba_region() {
-		let bgra = bgra_frame(4, 3, 16);
-		let mut patch = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_bgra_frame_loupe_patch_rgba(
-				4,
-				3,
-				16,
-				bgra.as_ptr(),
-				bgra.len(),
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 4.0, height: 3.0 },
-				0.0,
-				2.0,
-				3,
-				&mut patch,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(patch.width, 3);
-		assert_eq!(patch.height, 3);
-		assert_eq!(patch.len, 36);
-
-		let bytes = unsafe { slice::from_raw_parts(patch.rgba, patch.len) };
-
-		assert_eq!(&bytes[..8], &[10, 20, 30, 200, 10, 20, 30, 200]);
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut patch);
-		}
-	}
-
-	#[test]
-	fn ffi_bgra_frame_loupe_patch_rejects_invalid_storage() {
-		let bgra = bgra_frame(4, 3, 16);
-		let mut patch = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_bgra_frame_loupe_patch_rgba(
-				4,
-				3,
-				12,
-				bgra.as_ptr(),
-				bgra.len(),
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 4.0, height: 3.0 },
-				0.0,
-				2.0,
-				3,
-				&mut patch,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::InvalidInput);
-		assert!(patch.rgba.is_null());
-	}
-
-	#[test]
-	fn ffi_capture_frame_plan_returns_core_geometry() {
-		let mut plan = RsnapCaptureFramePlan::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_plan(
-				320,
-				180,
-				2.0,
-				RsnapCaptureFrameSourceKind::Window,
-				&mut plan,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(plan.canvas_width, 416.0);
-		assert_eq!(plan.canvas_height, 276.0);
-		assert_eq!(
-			plan.image_rect,
-			RsnapFloatRect { x: 48.0, y: 48.0, width: 320.0, height: 180.0 }
-		);
-		assert_eq!(plan.corner_radius, 9.9);
-		assert_eq!(plan.shadows[0].blur, 80.0);
-		assert_eq!(plan.shadows[1].offset_y, -22.0);
-	}
-
-	#[test]
-	fn ffi_capture_frame_aspect_fill_crop_rect_returns_core_rect() {
-		let mut rect = RsnapFloatRect::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_aspect_fill_crop_rect(
-				1_600, 900, 1_000.0, 1_000.0, &mut rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(rect, RsnapFloatRect { x: 350.0, y: 0.0, width: 900.0, height: 900.0 });
-	}
-
-	#[test]
-	fn ffi_capture_frame_background_plan_returns_core_preset() {
-		let mut plan = RsnapCaptureFrameBackgroundPlan::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_background_plan(
-				RsnapCaptureFrameBackgroundKind::Graphite,
-				&mut plan,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(plan.prefers_wallpaper, 0);
-		assert_eq!(plan.wallpaper_overlay_alpha, 0.0);
-		assert_eq!(plan.locations, [0.0, 0.54, 1.0]);
-		assert_eq!(
-			plan.colors[0],
-			RsnapCaptureFrameColorStop { red: 0.08, green: 0.09, blue: 0.11, alpha: 1.0 }
-		);
-		assert_eq!(
-			plan.colors[2],
-			RsnapCaptureFrameColorStop { red: 0.56, green: 0.59, blue: 0.64, alpha: 1.0 }
-		);
-	}
-
-	#[test]
-	fn ffi_capture_frame_wallpaper_request_returns_core_thumbnail_policy() {
-		let mut request = RsnapCaptureFrameWallpaperRequest::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_wallpaper_request_plan(
-				RsnapCaptureFrameBackgroundKind::SystemWallpaper,
-				1_535.2,
-				996.0,
-				&mut request,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(request.target_pixel_size, 1_536);
-		assert_eq!(request.overlay_alpha, 0.10);
-	}
-
-	#[test]
-	fn ffi_capture_frame_wallpaper_request_returns_empty_for_gradient_background() {
-		let mut request = RsnapCaptureFrameWallpaperRequest::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_wallpaper_request_plan(
-				RsnapCaptureFrameBackgroundKind::Aurora,
-				1_536.0,
-				996.0,
-				&mut request,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Empty);
-	}
-
-	#[test]
-	fn ffi_capture_frame_wallpaper_png_thumbnail_returns_owned_region() {
-		let path_buf =
-			env::temp_dir().join(format!("rsnap-ffi-wallpaper-thumb-{}-rgba.png", process::id()));
-		let png = rsnap_capture_core::RgbaExportImage::from_raw(
-			4,
-			2,
-			vec![
-				255, 0, 0, 255, 0, 255, 0, 255, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 0,
-				255, 0, 255, 255, 255, 255, 0, 255, 255, 20, 30, 40, 255,
-			],
-		)
-		.expect("test RGBA payload should create an image")
-		.to_png_bytes()
-		.expect("test RGBA image should encode as PNG");
-
-		fs::write(&path_buf, png).expect("test PNG should be written");
-
-		let path = CString::new(path_buf.to_string_lossy().as_bytes())
-			.expect("test PNG path should not contain interior NUL bytes");
-		let mut thumbnail = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_wallpaper_png_thumbnail(path.as_ptr(), 64, &mut thumbnail)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert!(thumbnail.width <= 64);
-		assert!(thumbnail.height <= 64);
-		assert_eq!(thumbnail.len, thumbnail.width as usize * thumbnail.height as usize * 4);
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut thumbnail);
-		}
-
-		let _ = fs::remove_file(path_buf);
-	}
-
-	#[test]
-	fn ffi_capture_frame_render_rgba_returns_owned_composition() {
-		let rgba = [255; 4 * 2 * 4];
-		let mut rendered = RsnapOwnedRgbaRegion::default();
-		let status = unsafe {
-			crate::rsnap_capture_frame_render_rgba(
-				4,
-				2,
-				rgba.as_ptr(),
-				rgba.len(),
-				2.0,
-				RsnapCaptureFrameSourceKind::DragRegion,
-				RsnapCaptureFrameBackgroundKind::Aurora,
-				RsnapCaptureFrameRenderKind::WindowSnapshot,
-				ptr::null(),
-				&mut rendered,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(rendered.width, 100);
-		assert_eq!(rendered.height, 98);
-		assert_eq!(rendered.len, 100 * 98 * 4);
-
-		let bytes = unsafe { slice::from_raw_parts(rendered.rgba, rendered.len) };
-		let first_source_pixel = ((48 * rendered.width as usize) + 48) * 4;
-
-		assert_eq!(&bytes[first_source_pixel..first_source_pixel + 4], &[255, 255, 255, 255]);
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut rendered);
-		}
-	}
-
-	#[test]
-	fn ffi_scroll_minimap_plan_returns_core_geometry() {
-		let mut plan = RsnapScrollMinimapPlan::default();
-		let status = unsafe {
-			crate::rsnap_scroll_minimap_plan(
-				RsnapFloatRect { x: 100.0, y: 100.0, width: 100.0, height: 100.0 },
-				100.0,
-				200.0,
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 500.0, height: 500.0 },
-				96.0,
-				44.0,
-				10.0,
-				10.0,
-				3.0,
-				20.0,
-				100.0,
-				&mut plan,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(plan.frame, RsnapFloatRect { x: 210.0, y: 54.0, width: 96.0, height: 192.0 });
-		assert_eq!(
-			plan.image_frame,
-			RsnapFloatRect { x: 213.0, y: 57.0, width: 90.0, height: 186.0 }
-		);
-		assert_eq!(plan.has_viewport_frame, 1);
-		assert_eq!(
-			plan.viewport_frame,
-			RsnapFloatRect { x: 213.0, y: 131.4, width: 90.0, height: 93.0 }
-		);
-	}
-
-	#[test]
-	fn ffi_scroll_minimap_plan_returns_empty_when_too_tight() {
-		let mut plan = RsnapScrollMinimapPlan::default();
-		let status = unsafe {
-			crate::rsnap_scroll_minimap_plan(
-				RsnapFloatRect { x: 100.0, y: 100.0, width: 100.0, height: 100.0 },
-				100.0,
-				200.0,
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 230.0, height: 60.0 },
-				96.0,
-				44.0,
-				10.0,
-				10.0,
-				3.0,
-				20.0,
-				100.0,
-				&mut plan,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Empty);
-	}
-
-	#[test]
-	fn ffi_frozen_selection_transform_hit_test_returns_core_kind() {
-		let mut kind = RsnapFrozenSelectionTransformKind::Move;
-		let status = unsafe {
-			crate::rsnap_frozen_selection_transform_hit_test(
-				102.0,
-				238.0,
-				RsnapFloatRect { x: 100.0, y: 80.0, width: 240.0, height: 160.0 },
-				12.0,
-				4.0,
-				&mut kind,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(kind, RsnapFrozenSelectionTransformKind::ResizeTopLeft);
-	}
-
-	#[test]
-	fn ffi_frozen_selection_transform_rect_returns_core_rect() {
-		let mut rect = RsnapFloatRect::default();
-		let status = unsafe {
-			crate::rsnap_frozen_selection_transform_rect(
-				RsnapFrozenSelectionTransformKind::ResizeBottomRight,
-				RsnapFloatRect { x: 100.0, y: 80.0, width: 240.0, height: 160.0 },
-				RsnapFloatRect { x: 0.0, y: 0.0, width: 500.0, height: 400.0 },
-				340.0,
-				80.0,
-				50.0,
-				300.0,
-				12.0,
-				&mut rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(rect, RsnapFloatRect { x: 100.0, y: 228.0, width: 12.0, height: 12.0 });
-	}
-
-	#[test]
-	fn ffi_auto_center_content_bounds_returns_core_rect() {
-		let rgba = auto_center_frame(
-			100,
-			80,
-			Some(RsnapPixelRect { x: 30, y: 20, width: 24, height: 18 }),
-		);
-		let mut rect = RsnapPixelRect::default();
-		let status = unsafe {
-			crate::rsnap_auto_center_content_bounds_rgba(
-				100,
-				80,
-				rgba.as_ptr(),
-				rgba.len(),
-				&mut rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Ok);
-		assert_eq!(rect, RsnapPixelRect { x: 30, y: 20, width: 24, height: 18 });
-		assert_eq!(
-			crate::rsnap_auto_center_margin_balance_shift_points(30.0, 24.0, 100.0, 50.0),
-			-4.0
-		);
-	}
-
-	#[test]
-	fn ffi_auto_center_content_bounds_returns_empty_for_uniform_image() {
-		let rgba = auto_center_frame(100, 80, None);
-		let mut rect = RsnapPixelRect::default();
-		let status = unsafe {
-			crate::rsnap_auto_center_content_bounds_rgba(
-				100,
-				80,
-				rgba.as_ptr(),
-				rgba.len(),
-				&mut rect,
-			)
-		};
-
-		assert_eq!(status, RsnapStatus::Empty);
-	}
-
-	#[cfg(target_os = "macos")]
-	#[test]
-	fn ffi_scroll_session_observes_downward_frame_and_exports() {
-		let base = scroll_frame(16, 96, 0);
-		let moved = scroll_frame(16, 96, 24);
-		let handle =
-			unsafe { crate::rsnap_scroll_session_create(16, 96, base.as_ptr(), base.len(), 8) };
-
-		assert!(!handle.is_null());
-
-		let mut result = RsnapScrollObserveResult::default();
-		let observe_status = unsafe {
-			crate::rsnap_scroll_session_observe_downward_frame(
-				handle,
-				16,
-				96,
-				moved.as_ptr(),
-				moved.len(),
-				&mut result,
-			)
-		};
-
-		assert_eq!(observe_status, RsnapStatus::Ok);
-		assert_eq!(result.kind, RsnapScrollObserveOutcomeKind::Committed as u32);
-		assert_eq!(result.growth_rows, 24);
-		assert_eq!(result.export_width, 16);
-		assert_eq!(result.export_height, 120);
-		assert_eq!(result.current_viewport_top_y, 24);
-
-		let mut export = RsnapOwnedRgbaRegion::default();
-
-		assert_eq!(
-			unsafe { crate::rsnap_scroll_session_take_export_rgba(handle, &mut export) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(export.width, 16);
-		assert_eq!(export.height, 120);
-		assert_eq!(export.len, 16 * 120 * 4);
-
-		let mut preview = RsnapOwnedRgbaRegion::default();
-
-		assert_eq!(
-			unsafe { crate::rsnap_scroll_session_take_preview_rgba(handle, &mut preview) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(preview.width, 8);
-		assert_eq!(preview.height, 60);
-		assert_eq!(preview.len, 8 * 60 * 4);
-
-		unsafe {
-			crate::rsnap_owned_rgba_region_release(&mut export);
-			crate::rsnap_owned_rgba_region_release(&mut preview);
-			crate::rsnap_scroll_session_destroy(handle);
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	#[test]
-	fn ffi_scroll_session_blocks_rewind_until_frontier_is_reacquired() {
-		let base = scroll_frame(16, 128, 0);
-		let first = scroll_frame(16, 128, 48);
-		let rewind = scroll_frame(16, 128, 24);
-		let below_frontier = scroll_frame(16, 128, 36);
-		let reacquired = scroll_frame(16, 128, 48);
-		let beyond_frontier = scroll_frame(16, 128, 60);
-		let handle =
-			unsafe { crate::rsnap_scroll_session_create(16, 128, base.as_ptr(), base.len(), 16) };
-
-		assert!(!handle.is_null());
-
-		let mut result = RsnapScrollObserveResult::default();
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_scroll_session_observe_downward_frame(
-					handle,
-					16,
-					128,
-					first.as_ptr(),
-					first.len(),
-					&mut result,
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(result.kind, RsnapScrollObserveOutcomeKind::Committed as u32);
-		assert_eq!(result.current_viewport_top_y, 48);
-		assert_eq!(result.export_height, 176);
-
-		for frame in [&rewind, &below_frontier, &reacquired] {
-			assert_eq!(
-				unsafe {
-					crate::rsnap_scroll_session_observe_downward_frame(
-						handle,
-						16,
-						128,
-						frame.as_ptr(),
-						frame.len(),
-						&mut result,
-					)
-				},
-				RsnapStatus::Ok
-			);
-			assert_ne!(result.kind, RsnapScrollObserveOutcomeKind::Committed as u32);
-			assert_eq!(result.current_viewport_top_y, 48);
-			assert_eq!(result.export_height, 176);
-		}
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_scroll_session_observe_downward_frame(
-					handle,
-					16,
-					128,
-					beyond_frontier.as_ptr(),
-					beyond_frontier.len(),
-					&mut result,
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(result.kind, RsnapScrollObserveOutcomeKind::Committed as u32);
-		assert_eq!(result.growth_rows, 12);
-		assert_eq!(result.current_viewport_top_y, 60);
-		assert_eq!(result.export_height, 188);
-
-		unsafe {
-			crate::rsnap_scroll_session_destroy(handle);
-		}
-	}
-
-	#[test]
-	fn ffi_click_freeze_request_carries_fixed_selection_payload() {
-		let handle = unsafe { crate::rsnap_session_create(default_config()) };
-		let mut request =
-			RsnapHostRequestValue { kind: u32::MAX, ..RsnapHostRequestValue::default() };
-
-		assert_eq!(unsafe { crate::rsnap_session_enter_live(handle) }, RsnapStatus::Ok);
-
-		let _ = unsafe { crate::rsnap_session_take_next_request(handle, &mut request) };
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_session_handle_host_event(
-					handle,
-					RsnapHostEvent {
-						kind: RsnapHostEventKind::PrimaryInteractionCompleted as u32,
-						point: RsnapPoint { x: 80, y: 110 },
-						has_point: 1,
-						rgb: RsnapRgb::default(),
-						has_rgb: 0,
-						active_monitor: RsnapMonitorRect {
-							id: 9,
-							origin: RsnapPoint { x: 0, y: 0 },
-							width: 1_440,
-							height: 900,
-							scale_factor_x1000: 2_000,
-						},
-						has_active_monitor: 1,
-						highlighted_window: RsnapWindowRect {
-							window_id: 42,
-							has_window_id: 1,
-							x: 20,
-							y: 30,
-							width: 60,
-							height: 80,
-						},
-						has_highlighted_window: 1,
-						toolbar_item_kind: 0,
-					},
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(
-			unsafe { crate::rsnap_session_take_next_request(handle, &mut request) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(request.kind, RsnapHostRequestKind::RequestFreezeSnapshot as u32);
-		assert_eq!(request.has_selection, 1);
-		assert_eq!(request.selection, RsnapRect { x: 20, y: 30, width: 60, height: 80 });
-		assert_eq!(request.selection_editable, 0);
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn ffi_drag_freeze_request_carries_editable_selection_payload() {
-		let handle = unsafe { crate::rsnap_session_create(default_config()) };
-		let mut request =
-			RsnapHostRequestValue { kind: u32::MAX, ..RsnapHostRequestValue::default() };
-
-		assert_eq!(unsafe { crate::rsnap_session_enter_live(handle) }, RsnapStatus::Ok);
-
-		let _ = unsafe { crate::rsnap_session_take_next_request(handle, &mut request) };
-
-		assert_eq!(
-			unsafe {
-				crate::rsnap_session_handle_host_event(
-					handle,
-					RsnapHostEvent {
-						kind: RsnapHostEventKind::PrimaryInteractionStarted as u32,
-						point: RsnapPoint { x: 80, y: 110 },
-						has_point: 1,
-						rgb: RsnapRgb::default(),
-						has_rgb: 0,
-						active_monitor: RsnapMonitorRect {
-							id: 9,
-							origin: RsnapPoint { x: 0, y: 0 },
-							width: 1_440,
-							height: 900,
-							scale_factor_x1000: 2_000,
-						},
-						has_active_monitor: 1,
-						highlighted_window: RsnapWindowRect {
-							window_id: 42,
-							has_window_id: 1,
-							x: 20,
-							y: 30,
-							width: 60,
-							height: 80,
-						},
-						has_highlighted_window: 1,
-						toolbar_item_kind: 0,
-					},
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(
-			unsafe {
-				crate::rsnap_session_handle_host_event(
-					handle,
-					RsnapHostEvent {
-						kind: RsnapHostEventKind::PrimaryInteractionCompleted as u32,
-						point: RsnapPoint { x: 140, y: 190 },
-						has_point: 1,
-						rgb: RsnapRgb::default(),
-						has_rgb: 0,
-						active_monitor: RsnapMonitorRect {
-							id: 9,
-							origin: RsnapPoint { x: 0, y: 0 },
-							width: 1_440,
-							height: 900,
-							scale_factor_x1000: 2_000,
-						},
-						has_active_monitor: 1,
-						highlighted_window: RsnapWindowRect {
-							window_id: 42,
-							has_window_id: 1,
-							x: 20,
-							y: 30,
-							width: 60,
-							height: 80,
-						},
-						has_highlighted_window: 1,
-						toolbar_item_kind: 0,
-					},
-				)
-			},
-			RsnapStatus::Ok
-		);
-		assert_eq!(
-			unsafe { crate::rsnap_session_take_next_request(handle, &mut request) },
-			RsnapStatus::Ok
-		);
-		assert_eq!(request.kind, RsnapHostRequestKind::RequestFreezeSnapshot as u32);
-		assert_eq!(request.has_selection, 1);
-		assert_eq!(request.selection, RsnapRect { x: 80, y: 110, width: 60, height: 80 });
-		assert_eq!(request.selection_editable, 1);
-
-		unsafe { crate::rsnap_session_destroy(handle) };
-	}
-
-	#[test]
-	fn abi_version_matches_constant() {
-		assert_eq!(crate::rsnap_host_ffi_abi_version(), RSNAP_HOST_FFI_ABI_VERSION);
-	}
-
-	fn auto_center_frame(width: u32, height: u32, content: Option<RsnapPixelRect>) -> Vec<u8> {
-		let mut rgba = vec![180_u8; (width * height * 4) as usize];
-
-		for pixel in rgba.chunks_exact_mut(4) {
-			pixel[3] = 255;
-		}
-
-		if let Some(content) = content {
-			for y in content.y..content.y + content.height {
-				for x in content.x..content.x + content.width {
-					let offset = ((y * width + x) * 4) as usize;
-
-					rgba[offset] = 24;
-					rgba[offset + 1] = 32;
-					rgba[offset + 2] = 40;
-				}
-			}
-		}
-
-		rgba
-	}
-
-	fn bgra_frame(width: u32, height: u32, bytes_per_row: usize) -> Vec<u8> {
-		let mut bytes = vec![0xEE; bytes_per_row * height as usize];
-
-		for y in 0..height {
-			for x in 0..width {
-				let offset = y as usize * bytes_per_row + x as usize * 4;
-
-				bytes[offset] = 30 + y as u8 * 15 + x as u8;
-				bytes[offset + 1] = 20 + y as u8 * 10 + x as u8;
-				bytes[offset + 2] = 10 + y as u8 * 5 + x as u8;
-				bytes[offset + 3] = 200 + y as u8 + x as u8;
-			}
-		}
-
-		bytes
-	}
-}
+mod tests;
